Created classes to interact with a display and to display screens
The DisplayHandler class provides the basic functions required for
interacting with a display.

There is currently an implementation that communicates with ILI9341 display
using the "Adafruit_ILI9341" library.

The Screen class is a simple base class for screens that will be displayed on the display.
It currently contains two functions:
    1. `prepare`: Clears the display and calls the sub-classes' prepare implementation.
    2. `dispose`: Currently only calling the sub-classes' dispose implementation.

`SetupScreen` is a simple implementation of screen whose purpose is to show the initialization steps when
the board starts.