Added temperatures viewing classes
`TemperaturesScreen` is the screen that shows the temperatures on the display
The `showTemperatures` method accepts the vector of `TemperatureReading` generated
by the `ITemperatureSensor::getReadings` method.
By default the number of columns is 2.

`AddressMappingTemperatureSensor` is a decorator designed to map the sensor's address
to a more human-readable label based on a given mapping (if there is not a label for the address, the address would be used).