Added Firebase authentication for MQTT connection
Previously the user's id was provided in a config file, and connection to the MQTT broker was based on the user's id and device's id.

To improve security, connection to the MQTT broker now requires an id token as a password.

The new `FirebaseAuthHandler` class handles the connection and communication with Firebase Auth to provide both an id token and the user's id.

The configuration files now also contain a loading function for the Firebase config file which provides the email, password and api-key of the project.

And the `MqttHandler` now gets the user's id from Firebase and uses the generated token to connect.